requirements.txt
streamlit==1.37.1
pyautogen==0.2.20
google-generativeai==0.7.2
python-dotenv==0.21.0
Pillow==10.4.0
-e .